feat(stations): get the closing time for a station
Get the closing time for a station for Saturdays and Sundays as these
were the easiest to do initially.

Next, we need to add support for weekdays which will means we need to
determine if the specified day is within the `com:MondayToFriday` key as
that's what's present for Cardiff Central station.

Looking at other stations, though, the format within the XML for opening
times is consistent with separate entries for Saturday, Sunday, and
weekdays.